+++ lib/rest/Conversations.d.ts
@@ -8,7 +8,13 @@
 import Domain = require('../base/Domain');
 import Twilio = require('./Twilio');
 import V1 = require('./conversations/V1');
+import { ConfigurationListInstance } from './conversations/v1/configuration';
 import { ConversationListInstance } from './conversations/v1/conversation';
+import { CredentialListInstance } from './conversations/v1/credential';
+import { NotificationListInstance } from './conversations/v1/notification';
+import { RoleListInstance } from './conversations/v1/role';
+import { ServiceListInstance } from './conversations/v1/service';
+import { UserListInstance } from './conversations/v1/user';
 import { WebhookListInstance } from './conversations/v1/webhook';
 
 
@@ -20,7 +26,13 @@ declare class Conversations extends Domain {
    */
   constructor(twilio: Twilio);
 
+  readonly configuration: ConfigurationListInstance;
   readonly conversations: ConversationListInstance;
+  readonly credentials: CredentialListInstance;
+  readonly notifications: NotificationListInstance;
+  readonly roles: RoleListInstance;
+  readonly services: ServiceListInstance;
+  readonly users: UserListInstance;
   readonly v1: V1;
   readonly webhooks: WebhookListInstance;
 }