+++ lib/rest/Events.d.ts
@@ -8,6 +8,7 @@
 import Domain = require('../base/Domain');
 import Twilio = require('./Twilio');
 import V1 = require('./events/V1');
+import { EventTypeListInstance } from './events/v1/eventType';
 import { SinkListInstance } from './events/v1/sink';
 import { SubscriptionListInstance } from './events/v1/subscription';
 
@@ -20,6 +21,7 @@ declare class Events extends Domain {
    */
   constructor(twilio: Twilio);
 
+  readonly eventTypes: EventTypeListInstance;
   readonly sinks: SinkListInstance;
   readonly subscriptions: SubscriptionListInstance;
   readonly v1: V1;