+++ lib/rest/conversations/v1/conversation/message/deliveryReceipt.d.ts
@@ -19,7 +19,7 @@ type DeliveryReceiptDeliveryStatus = 'read'|'failed'|'delivered'|'undelivered'|'
  * Use them with caution.
  *
  * @param version - Version of the resource
- * @param conversationSid - The conversation_sid
+ * @param conversationSid - The unique id of the Conversation for this message.
  * @param messageSid - The sid of the message the delivery receipt belongs to
  */
 declare function DeliveryReceiptList(version: V1, conversationSid: string, messageSid: string): DeliveryReceiptListInstance;
@@ -198,6 +198,7 @@ interface DeliveryReceiptPayload extends DeliveryReceiptResource, Page.TwilioRes
 }
 
 interface DeliveryReceiptResource {
+  account_sid: string;
   channel_message_sid: string;
   conversation_sid: string;
   date_created: Date;
@@ -225,7 +226,7 @@ declare class DeliveryReceiptContext {
    *
    * @param version - Version of the resource
    * @param conversationSid - The unique id of the Conversation for this delivery receipt.
-   * @param messageSid - The sid of the message the delivery receipt belongs to
+   * @param messageSid - The sid of the message the delivery receipt belongs to.
    * @param sid - A 34 character string that uniquely identifies this resource.
    */
   constructor(version: V1, conversationSid: string, messageSid: string, sid: string);
@@ -252,13 +253,14 @@ declare class DeliveryReceiptInstance extends SerializableClass {
    *
    * @param version - Version of the resource
    * @param payload - The instance payload
-   * @param conversationSid - The conversation_sid
+   * @param conversationSid - The unique id of the Conversation for this message.
    * @param messageSid - The sid of the message the delivery receipt belongs to
    * @param sid - A 34 character string that uniquely identifies this resource.
    */
   constructor(version: V1, payload: DeliveryReceiptPayload, conversationSid: string, messageSid: string, sid: string);
 
   private _proxy: DeliveryReceiptContext;
+  accountSid: string;
   channelMessageSid: string;
   conversationSid: string;
   dateCreated: Date;