+++ lib/rest/serverless/v1/service/environment/variable.d.ts
@@ -19,7 +19,7 @@ import { SerializableClass } from '../../../../../interfaces';
  *
  * @param version - Version of the resource
  * @param serviceSid - The SID of the Service that the Variable resource is associated with
- * @param environmentSid - The SID of the environment in which the variable exists
+ * @param environmentSid - The SID of the Environment in which the Variable exists
  */
 declare function VariableList(version: V1, serviceSid: string, environmentSid: string): VariableListInstance;
 
@@ -27,7 +27,7 @@ declare function VariableList(version: V1, serviceSid: string, environmentSid: s
  * Options to pass to update
  *
  * @property key - A string by which the Variable resource can be referenced
- * @property value - A string that contains the actual value of the variable
+ * @property value - A string that contains the actual value of the Variable
  */
 interface VariableInstanceUpdateOptions {
   key?: string;
@@ -158,7 +158,7 @@ interface VariableListInstance {
  * Options to pass to create
  *
  * @property key - A string by which the Variable resource can be referenced
- * @property value - A string that contains the actual value of the variable
+ * @property value - A string that contains the actual value of the Variable
  */
 interface VariableListInstanceCreateOptions {
   key: string;
@@ -253,7 +253,7 @@ declare class VariableContext {
    *
    * @param version - Version of the resource
    * @param serviceSid - The SID of the Service to fetch the Variable resource from
-   * @param environmentSid - The SID of the environment with the Variable resource to fetch
+   * @param environmentSid - The SID of the Environment with the Variable resource to fetch
    * @param sid - The SID of the Variable resource to fetch
    */
   constructor(version: V1, serviceSid: string, environmentSid: string, sid: string);
@@ -301,7 +301,7 @@ declare class VariableInstance extends SerializableClass {
    * @param version - Version of the resource
    * @param payload - The instance payload
    * @param serviceSid - The SID of the Service that the Variable resource is associated with
-   * @param environmentSid - The SID of the environment in which the variable exists
+   * @param environmentSid - The SID of the Environment in which the Variable exists
    * @param sid - The SID of the Variable resource to fetch
    */
   constructor(version: V1, payload: VariablePayload, serviceSid: string, environmentSid: string, sid: string);